Added DynamicLoader to the engine #184.
DynamicLoader is the first version of a unified loader that works both with the tool and standalone.
It's "dynamic" in that it can update objects in the world based on new configs (data model format).

The first version covers all the functionality that was in the tool loaders, but is still missing:
* Skeletons and Animations
* Script components

It's written in coffeescript, which made it necessary to add a coffeescript source tree paralell to the
regular source tree. Use cake coffee to compile the coffeescript into the js source tree. cake minify
requires cake coffee.