Fix non-selectable debug meshes
The DebugRenderSystem keeps a cache of renderables, which are connected to the entities by id. This
cache lives on even if the entities are deleted. So if a user unloads a scene, and then loads it again,
the renderable cache will have renderables for all the scene entities, but they will have the wrong
index. Which makes them unselectable, as the picking system is based on indices.

Solution is to delete the cache when entities are deleted (shouldn't be that expensive).